update to use existing component, remove package-lock
